Fixes #1014
